refactor: generalize PaneHistory into shared MRUHistory<Element> (spec C §6)
Moves the MRU navigation history from Sources/TBDApp/Panes/PaneHistory.swift
into Sources/TBDShared/PanelSurface/MRUHistory.swift as a public generic
MRUHistory<Element: Codable & Equatable & Sendable>, with `PaneHistory` kept
as a typealias (MRUHistory<PaneContent>) so existing AppState call sites
compile unchanged. Same shipped PR #472 algorithm, only genericized.

The 17 existing PaneHistoryTests are split by target dependency rather than
moved wholesale: the 9 pure MRU-algorithm tests move to Tests/TBDSharedTests
(now testing the generic through the PaneHistory typealias) since TBDShared
must not depend on TBDApp/AppKit/SwiftUI; the 8 AppState-integration tests
stay in Tests/TBDAppTests/PaneHistoryAppStateTests.swift since AppState is
TBDApp-only and TBDSharedTests has no TBDApp dependency. Two new tests
pin the `seeded(with:)` API and the exact #472-persisted wire shape
({entries, cursor}) for decode compatibility.
